CHANGELOG.md,.github/: Add CHANGELOG process for monitoring stack
This makes it easier for us to track what we changed in the current
release, and saves us time when writing release notes and talking
with customer support or writing docs. This is informational only and
OpenShift release notes should serve as the official changelog.

Also adds a reminder to fill in the CHANGELOG as part of PR templates.